LSTM/src/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.1.0
torchvision==0.16.0
matplotlib==3.7.2
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2 